requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
defusedxml==0.7.1
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
zipp==3.6.0
